requirements.txt
pymongo
beautifulsoup4
Flask==2.2.2
Flask-Session==0.4.0
google-api-core==2.10.0
google-api-python-client==2.60.0
google-auth==2.11.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.4
Jinja2==3.1.2
numpy
pandas==1.4.4
python-dateutil==2.8.2
pytube==12.1.0
requests==2.28.1
urllib3
python-dotenv
seaborn
mysql-connector-python